requirements.txt
discord.py
aiohttp
mcrcon
requests
